requirements.txt
# This requirements file specifies the dependencies for the QuickCompress project.
# Minimum Python version required: 3.8

pillow==10.4.0
pillow-avif-plugin==1.4.6
pytest==8.3.3
pytest-qt==4.4.0
PyQt6==6.7.1
